organizing lib code
